+++ app/src/main/kotlin/com/hedvig/app/feature/claims/ui/ClaimsUtil.kt
@@ -5,11 +5,8 @@ import android.content.Intent
 import androidx.fragment.app.FragmentManager
 import com.hedvig.android.hanalytics.featureflags.FeatureManager
 import com.hedvig.android.hanalytics.featureflags.flags.Feature
-import com.hedvig.android.odyssey.OdysseyClaimsFlowActivity
 import com.hedvig.app.R
-import com.hedvig.android.odyssey.model.ItemType
 import com.hedvig.android.odyssey.search.SearchActivity
-import com.hedvig.android.odyssey.search.SearchActivity.Companion.ODYSSEY_URL
 import com.hedvig.app.feature.claims.ui.pledge.HonestyPledgeBottomSheet
 
 suspend fun startClaimsFlow(
@@ -21,8 +18,7 @@ suspend fun startClaimsFlow(
 ) {
   if (featureManager.isFeatureEnabled(Feature.USE_ODYSSEY_CLAIM_FLOW)) {
 
-    val intent = Intent(context, SearchActivity::class.java)
-      .putExtra(ODYSSEY_URL, context.getString(R.string.ODYSSEY_URL))
+    val intent = SearchActivity.newInstance(context, context.getString(R.string.ODYSSEY_URL))
 
     if (registerForResult != null) {
       registerForResult(intent)
