--- feature-odyssey/src/main/kotlin/com/hedvig/android/odyssey/search/GetClaimEntryPoints.kt
@@ -1,12 +1,0 @@
-package com.hedvig.android.odyssey.search
-
-import com.hedvig.android.odyssey.model.SearchableClaim
-
-interface GetClaimEntryPoints {
-  suspend operator fun invoke(): CommonClaimsResult
-}
-
-sealed interface CommonClaimsResult {
-  data class Error(val message: String) : CommonClaimsResult
-  data class Success(val searchableClaims: List<SearchableClaim>) : CommonClaimsResult
-}
