+++ feature-odyssey/src/main/kotlin/com/hedvig/android/odyssey/search/SearchActivity.kt
@@ -1,9 +1,10 @@
 package com.hedvig.android.odyssey.search
 
+import android.content.Context
+import android.content.Intent
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
@@ -12,13 +13,10 @@ import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.platform.LocalSoftwareKeyboardController
 import androidx.compose.ui.unit.dp
 import com.hedvig.android.auth.android.AuthenticatedObserver
 import com.hedvig.android.core.designsystem.theme.HedvigTheme
@@ -27,6 +25,10 @@ import com.hedvig.android.odyssey.OdysseyClaimsFlowActivity
 import org.koin.androidx.viewmodel.ext.android.getViewModel
 
 class SearchActivity : ComponentActivity() {
+
+  private val odysseyUrl by lazy {
+    intent.getStringExtra(ODYSSEY_URL) ?: throw java.lang.IllegalArgumentException("No url found")
+  }
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
@@ -37,8 +39,10 @@ class SearchActivity : ComponentActivity() {
       val viewState by viewModel.viewState.collectAsState()
 
       val commonClaimId = viewState.selectedClaim?.id
-      if (commonClaimId != null) {
-        startClaimsFlow(viewModel, commonClaimId)
+      LaunchedEffect(commonClaimId) {
+        if (commonClaimId != null) {
+          startClaimsFlow(viewModel, commonClaimId)
+        }
       }
 
       HedvigTheme {
@@ -79,7 +83,7 @@ class SearchActivity : ComponentActivity() {
   private fun startClaimsFlow(viewModel: SearchViewModel, commonClaimId: String?) {
     val intent = OdysseyClaimsFlowActivity.newInstance(
       context = this,
-      odysseyUrl = intent.getStringExtra(ODYSSEY_URL) ?: throw java.lang.IllegalArgumentException("No url found"),
+      odysseyUrl = odysseyUrl,
       commonClaimId = commonClaimId,
     )
     startActivity(intent)
@@ -88,5 +92,8 @@ class SearchActivity : ComponentActivity() {
 
   companion object {
     const val ODYSSEY_URL = "ODYSSEY_URL_EXTRA"
+
+    fun newInstance(context: Context, url: String) = Intent(context, SearchActivity::class.java)
+      .putExtra(ODYSSEY_URL, url)
   }
 }
